Added files via upload
So, I fixed the  issue with not receiving the right answer when putting in the yes/no responses. However this would not be a permanent fix, because it is only for  the  files and the arrays given in the project. I have all the files necessary within the tich file. Cheers!